refactor(worker-agent): split delivery eval fixtures
